+++ PlotsBase/src/Axes.jl
@@ -1,7 +1,7 @@
 module Axes
 
 export Axis, Extrema, tickfont, guidefont, widen_factor, scale_inverse_scale_func
-export sort_3d_axes, axes_letters, process_axis_arg!, has_ticks, get_axis
+export sort_3d_axes, axes_letters, process_axis_arg!, has_ticks, get_axis, get_guide
 
 import ..PlotsBase
 import ..PlotsBase: Subplot, DefaultsDict, attr!
@@ -327,6 +327,14 @@ function PlotsBase.attr!(axis::Axis, args...; kw...)
             foreach(x -> discrete_value!(axis, x), v)  # add these discrete values to the axis
         elseif k ≡ :lims && isa(v, NTuple{2,Dates.TimeType})
             plotattributes[k] = (Dates.value(v[1]), Dates.value(v[2]))
+        elseif k ≡ :guide && v isa AbstractString && isempty(v)
+            plotattributes[:unitformat] = :nounit
+            plotattributes[k] = v
+        elseif k ≡ :unit
+            if !isnothing(plotattributes[k]) && plotattributes[k] != v
+                @warn "Overriding unit for $(axis[:letter]) axis: $(plotattributes[k]) -> $v.  This will produce a plot, but series plotted before the override cannot update and will therefore be incorrectly treated as if they had the new units."
+            end
+            plotattributes[k] = v
         else
             plotattributes[k] = v
         end
@@ -339,6 +347,53 @@ function PlotsBase.attr!(axis::Axis, args...; kw...)
 
     axis
 end
+
+function get_guide(axis::Axis)
+    if isnothing(axis[:unit]) || axis[:unitformat] == :nounit
+        return axis[:guide]
+    else
+        ustr = if PlotsBase.backend_name() ≡ :pgfplotsx
+            pgext = Base.get_extension(PlotsBase, :PGFPlotsXExt)
+            isnothing(pgext) && error("PGFPlotsX extension not loaded. Problems")
+            pgext.Latexify.latexify(axis[:unit])
+        else
+            string(axis[:unit])
+        end
+        if isempty(axis[:guide])
+            return ustr
+        end
+        return format_unit_label(axis[:guide], ustr, axis[:unitformat])
+    end
+end
+
+# Keyword options for unit formats
+const UNIT_FORMATS = Dict(
+    :round => ('(', ')'),
+    :square => ('[', ']'),
+    :curly => ('{', '}'),
+    :angle => ('<', '>'),
+    :slash => '/',
+    :space => " ",
+    :slashround => (" / (", ")"),
+    :slashsquare => (" / [", "]"),
+    :slashcurly => (" / {", "}"),
+    :slashangle => (" / <", ">"),
+    :verbose => " in units of ",
+    :none => nothing,
+    :nounit => (l, u)->l, # no unit, just label
+)
+
+# All options for unit formats
+format_unit_label(l, u, f::Nothing)                    = l
+format_unit_label(l, u, f::Function)                   = f(l, u)
+format_unit_label(l, u, f::AbstractString)             = string(l, f, u)
+format_unit_label(l, u, f::NTuple{2,<:AbstractString}) = string(l, f[1], u, f[2])
+format_unit_label(l, u, f::NTuple{3,<:AbstractString}) = string(f[1], l, f[2], u, f[3])
+format_unit_label(l, u, f::Char)                       = string(l, ' ', f, ' ', u)
+format_unit_label(l, u, f::NTuple{2,Char})             = string(l, ' ', f[1], u, f[2])
+format_unit_label(l, u, f::NTuple{3,Char})             = string(f[1], l, ' ', f[2], u, f[3])
+format_unit_label(l, u, f::Bool)                       = f ? format_unit_label(l, u, :round) : format_unit_label(l, u, nothing)
+format_unit_label(l, u, f::Symbol)                     = format_unit_label(l, u, UNIT_FORMATS[f])
 
 # -----------------------------------------------------------------------------