+++ PlotsBase/src/utils.jl
@@ -931,26 +931,6 @@ end
 _generate_doclist(attributes) =
     replace(join(sort(collect(attributes)), "\n- "), "_" => "\\_")
 
-# for UnitfulExt - cannot reside in `UnitfulExt` (macro)
-function protectedstring end  # COV_EXCL_LINE
-
-"""
-    P_str(s)
-
-(Unitful extension only).
-Creates a string that will be Protected from recipe passes.
-
-Example:
-```julia
-julia> using Unitful
-julia> plot([0,1]u"m", [1,2]u"m/s^2", xlabel=P"This label will NOT display units")
-julia> plot([0,1]u"m", [1,2]u"m/s^2", xlabel="This label will display units")
-```
-"""
-macro P_str(s)
-    protectedstring(s)
-end
-
 # for `PGFPlotsx` together with `UnitfulExt`
 function pgfx_sanitize_string end  # COV_EXCL_LINE